src/handlers/SshServerHandler.ts
import { Client, ConnectConfig, ClientChannel } from 'ssh2';
import { ServerConfig } from '../types/ServerConfig';
import { PaginatedResponse } from '../types/PaginatedResponse';
import { SystemInfo } from '../types/SystemInfo';
import { ServerHandlerInterface } from '../types/ServerHandlerInterface';

class SshServerHandler implements ServerHandlerInterface {
  private client: Client;
  private config: ServerConfig;
  private globalState: { [key: string]: any };

  constructor(config: ServerConfig) {
    this.client = new Client();
    this.config = config;
    this.globalState = {};
  }

  initializeSSHClient(): Client {
    return new Client();
  }

  async connect(): Promise<void> {
    this.client.connect({
      host: this.config.host,
      username: this.config.username,
      privateKey: this.config.privateKeyPath,
      readyTimeout: 20000,
    });
  }

  async executeCommand(command: string, timeout?: number, directory?: string): Promise<{ stdout: string; stderr: string }> {
    return new Promise((resolve, reject) => {
      let stdout = '';
      let stderr = '';
      this.client.exec(command, (err, stream) => {
        if (err) return reject(err);
        stream.on('data', (data: Buffer) => {
          stdout += data.toString();
        }).on('stderr', (data: Buffer) => {
          stderr += data.toString();
        }).on('close', (code: number) => { // Explicitly typed code parameter
          if (code === 0) {
            resolve({ stdout, stderr });
          } else {
            reject(new Error(`Command failed with exit code ${code}`));
          }
        });
      });
    });
  }

  // Implementing missing methods
  listFiles(directory: string): Promise<PaginatedResponse<any>> {
    throw new Error("Method not implemented.");
  }

  createFile(directory: string, filename: string, content: string, backup: boolean): Promise<boolean> {
    throw new Error("Method not implemented.");
  }

  updateFile(filePath: string, pattern: string, replacement: string, backup: boolean): Promise<boolean> {
    throw new Error("Method not implemented.");
  }

  amendFile(filePath: string, content: string): Promise<boolean> {
    throw new Error("Method not implemented.");
  }

  getSystemInfo(): Promise<SystemInfo> {
    throw new Error("Method not implemented.");
  }
}

export { SshServerHandler };
